PROD-2078 #comment make the contact support form a top-level lib item and have the modal use it #time 30m
