+++ src-ts/lib/contact-support/ContactSupportForm.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/includes';
+@import '../styles/includes';
 
 .contact-support-intro {
