+++ src-ts/lib/contact-support/index.ts
@@ -1,0 +1,2 @@
+export { default as ContactSupportForm } from './ContactSupportForm'
+export { contactSupportFormDef } from './contact-support-form.config'
